fix(tests): Fix issue where circular dependencies would make tests not run
