+++ src/config.ts
@@ -1,12 +1,5 @@
 import { config } from 'dotenv';
-import EvmNetwork from './networks/evm/EvmNetwork';
 import packageJson from '../package.json';
-import { LayerZeroModule } from './modules/layerZero/LayerZeroModule';
-import { PushPairModule } from './modules/pushPair/PushPairModule';
-import { BalanceCheckerModule } from './modules/balanceChecker/BalanceCheckerModule';
-import { PairCheckerModule } from './modules/pairChecker/PairCheckerModule'
-import { FetchJob } from './jobs/fetch/FetchJob';
-import { NearNetwork } from './networks/near/NearNetwork';
 
 config();
 
@@ -18,9 +11,6 @@ export const PROJECT_NAME = packageJson.name;
 export const PROJECT_VERSION = packageJson.version;
 export const APP_CONFIG_LOCATION = process.env.APP_CONFIG_LOCATION ?? './config.json'
 
-export const AVAILABLE_NETWORKS = [EvmNetwork, NearNetwork];
-export const AVAILABLE_MODULES = [LayerZeroModule, PushPairModule, BalanceCheckerModule, PairCheckerModule];
-export const AVAILABLE_JOBS = [FetchJob];
 export const NODE_ID = process.env.NODE_ID ?? 'Anonymous';
 
 export const ENABLE_TELEGRAM_NOTIFICATIONS = process.env.ENABLE_TELEGRAM_NOTIFICATIONS === 'true';
