+++ src/models/DataRequest.ts
@@ -1,6 +1,6 @@
 import Big from "big.js";
 import { Block } from "./Block";
-import { Network } from "./Network";
+import { INetwork } from "./INetwork";
 import { Outcome } from "./Outcome";
 import { TxCallParams } from "./TxCallParams";
 
@@ -8,13 +8,12 @@ export interface CreatedInfo {
     block: Block;
 }
 
-
 export interface DataRequest {
     internalId: string;
     args: string[];
     createdInfo: CreatedInfo;
-    originNetwork: Network;
-    targetNetwork: Network;
+    originNetwork: INetwork;
+    targetNetwork: INetwork;
     confirmationsRequired: Big;
     extraInfo: {
         [key: string]: any;
