+++ src/models/DataRequestBatch.test.ts
@@ -7,7 +7,7 @@ import { hasBatchEnoughConfirmations } from "./DataRequestBatch";
 describe('DataRequestBatch', () => {
     describe('hasBatchEnoughConfirmations', () => {
         it('should return false when one needs more confirmations required', () => {
-            const isConfirmed = hasBatchEnoughConfirmations({
+            const details = hasBatchEnoughConfirmations({
                 internalId: '1',
                 targetNetwork: createNetworkMock(),
                 requests: [
@@ -20,11 +20,11 @@ describe('DataRequestBatch', () => {
                 ],
             }, createBlockMock({ number: new Big(10) }));
 
-            expect(isConfirmed).toBe(false);
+            expect(details.confirmed).toBe(false);
         });
 
         it('should return true when all requests are confirmed', () => {
-            const isConfirmed = hasBatchEnoughConfirmations({
+            const details = hasBatchEnoughConfirmations({
                 internalId: '1',
                 targetNetwork: createNetworkMock(),
                 requests: [
@@ -43,11 +43,11 @@ describe('DataRequestBatch', () => {
                 ],
             }, createBlockMock({ number: new Big(150) }));
 
-            expect(isConfirmed).toBe(true);
+            expect(details.confirmed).toBe(true);
         });
 
         it('should return false when one request is not confirmed yet', () => {
-            const isConfirmed = hasBatchEnoughConfirmations({
+            const details = hasBatchEnoughConfirmations({
                 internalId: '1',
                 targetNetwork: createNetworkMock(),
                 requests: [
@@ -66,7 +66,7 @@ describe('DataRequestBatch', () => {
                 ],
             }, createBlockMock({ number: new Big(150) }));
 
-            expect(isConfirmed).toBe(false);
+            expect(details.confirmed).toBe(false);
         });
     });
 });
