+++ src/models/Job.ts
@@ -1,8 +1,9 @@
 import { AppConfig } from "./AppConfig";
 import { DataRequest } from "./DataRequest";
+import { IJob } from "./IJob";
 import { Outcome } from "./Outcome";
 
-export class Job {
+export class Job implements IJob {
     static type = "job";
     id: string = 'job';
     type: string = Job.type;
